+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## All Changes
 
+- 2026-08-13 feat(website): expand commercial page content (#224) - Eduardo A.
+- 2026-08-13 feat(website): expand commercial page content - Eduardo A.
 - 2026-08-13 feat(website): add white Jumentix site icon (#222) - Eduardo A.
 - 2026-08-13 feat(website): add white Jumentix site icon - Eduardo A.
 - 2026-08-13 feat(website): show energized Jumentix eating sugarcane (#221) - Eduardo A.